+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## [2025.10.23]
+
+### Added
+
+- Designed kernel and enhanced interfaces for the StageLight component
+
+### Updated
+
+- Changed the name of the StageLight.java file from the previous proof of concept assignment to StageLightProofOfConcept.java
+
+```
+
 ## [2025.10.09]
 
 ### Added